add basic test for db

+++ app/models/Friendship.scala
@@ -1,6 +1,6 @@
 package models
 
-case class Friendship(id1: Long, id2: Long, status: Int, actionId: Int)
+case class Friendship(id1: Long, id2: Long, status: Int, actionId: Long)
 
 object Friendship {
   val STATUS_PENDING = 0
